Rename the train_test_split module
To avoid naming conflict with the function
