+++ ethicml/__init__.py
@@ -52,7 +52,7 @@ from .preprocessing.biased_split import *
 from .preprocessing.domain_adaptation import *
 from .preprocessing.feature_binning import *
 from .preprocessing.scaling import *
-from .preprocessing.train_test_split import *
+from .preprocessing.split import *
 from .utility.activation import *
 from .utility.data_helpers import *
 from .utility.data_structures import *
