+++ ethicml/preprocessing/__init__.py
@@ -5,4 +5,4 @@ from .biased_split import *
 from .domain_adaptation import *
 from .feature_binning import *
 from .scaling import *
-from .train_test_split import *
+from .split import *
